+++ src/test/java/config/fixtures/subway/StationLineMockData.java
@@ -5,14 +5,14 @@ import subway.dto.StationLineRequest;
 public class StationLineMockData {
 
     public static final StationLineRequest 신분당선 =
-            new StationLineRequest("신분당선", "bg-red-600", 1, 2, 10);
+            new StationLineRequest("신분당선", "bg-red-600", 1L, 2L, 10);
     public static final StationLineRequest 분당선 =
-            new StationLineRequest("신분당선", "bg-green-600", 2, 4, 20);
+            new StationLineRequest("신분당선", "bg-green-600", 2L, 4L, 20);
 
     public static final StationLineRequest 수정된_신분당선 =
-            new StationLineRequest("수정된_신분당선", "bg-blue-100", 1, 2, 10);
+            new StationLineRequest("수정된_신분당선", "bg-blue-100", 1L, 2L, 10);
 
-    public static StationLineRequest 호남선_생성_상행_하행_설정(int upStationId, int downStationId) {
+    public static StationLineRequest 호남선_생성_상행_하행_설정(Long upStationId, Long downStationId) {
         return new StationLineRequest("호남선", "bg-blue-100", upStationId, downStationId, 10);
     };
 }
